# docs/swagger.yaml
definitions:
  models.AuthSwag:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  models.Menu:
    properties:
      created_on:
        type: integer
      deleted_on:
        type: integer
      id:
        type: integer
      method:
        type: string
      modified_on:
        type: integer
      name:
        type: string
      path:
        type: string
      type:
        type: string
    type: object
  models.Role:
    properties:
      created_on:
        type: integer
      deleted_on:
        type: integer
      id:
        type: integer
      menu:
        items:
          $ref: '#/definitions/models.Menu'
        type: array
      modified_on:
        type: integer
      name:
        type: string
    type: object
  models.User:
    properties:
      created_on:
        type: integer
      deleted_on:
        type: integer
      id:
        type: integer
      modified_on:
        type: integer
      password:
        type: string
      role:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      username:
        type: string
    type: object
host: 127.0.0.1:8000
info:
  contact:
    email: hequan2011@sina.com
    name: hequan
    url: https://github.com/hequan2017
  description: go-admin
  license:
    name: MIT
    url: https://github.com/hequan2017/go-admin/blob/master/LICENSE
  termsOfService: https://github.com/hequan2017/go-admin
  title: go-admin
  version: "2.0"
paths:
  /api/v1/menus:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: '{ "code": 200, "data": {}, "msg": "ok" }'
          schema:
            type: string
      summary: 获取所有菜单
      tags:
      - menu
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Menu'
      produces:
      - application/json
      responses:
        "200":
          description: '{ "code": 200, "data": {}, "msg": "ok" }'
          schema:
            type: string
      summary: 增加菜单
      tags:
      - menu
  /api/v1/menus/:id:
    delete:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{ "code": 200, "data": {}, "msg": "ok" }'
          schema:
            type: string
      summary: 删除菜单
      tags:
      - menu
    put:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Menu'
      produces:
      - application/json
      responses:
        "200":
          description: '{ "code": 200, "data": {}, "msg": "ok" }'
          schema:
            type: string
      summary: 更新菜单
      tags:
      - menu
  /api/v1/roles:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: '{ "code": 200, "data": {}, "msg": "ok" }'
          schema:
            type: string
      summary: 获取所有角色
      tags:
      - role
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Role'
      produces:
      - application/json
      responses:
        "200":
          description: '{ "code": 200, "data": {}, "msg": "ok" }'
          schema:
            type: string
      summary: 增加角色
      tags:
      - role
  /api/v1/roles/:id:
    delete:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{ "code": 200, "data": {}, "msg": "ok" }'
          schema:
            type: string
      summary: 删除角色
      tags:
      - role
    put:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Role'
      produces:
      - application/json
      responses:
        "200":
          description: '{ "code": 200, "data": {}, "msg": "ok" }'
          schema:
            type: string
      summary: 更新角色
      tags:
      - role
  /api/v1/userInfo:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: '{ "code": 200, "data": {"lists":""}, "msg": "ok" }'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 获取登录用户信息
      tags:
      - users
  /api/v1/users:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: '{ "code": 200, "data": {}, "msg": "ok" }'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 获取所有用户
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: '{ "code": 200, "data": {}, "msg": "ok" }'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 增加用户
      tags:
      - users
  /api/v1/users/:id:
    delete:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{ "code": 200, "data": {}, "msg": "ok" }'
          schema:
            type: string
      summary: 删除用户
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: '{ "code": 200, "data": {}, "msg": "ok" }'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 更新用户
      tags:
      - users
  /auth:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.AuthSwag'
      produces:
      - application/json
      responses:
        "200":
          description: '{ "code": 200, "data": {}, "msg": "ok" }'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 获取登录token 信息
      tags:
      - auth
swagger: "2.0"